use empty html description if no summary is available
when using an empty description `IllegalStateException` is thrown with
the message "One of HTML description or Markdown description must be
defined for rule [repository=OCLint, key=high cyclomatic complexity]".

it would seem that the description is required, and since the process
to update the oclint rules does not yet include a proper description,
the description would be essentially empty.