add support for test of categories
tests for class categories use the "BaseClassName+CategoryNameTest.m"
format as the filename, however the actual test class within use the
"BaseClassName_CategoryNameTest" format (since `+` is illegal to use
within identifier names).

when reading the test suites from a report we have to replace the `_`
with `+` in order to get the class filename.