improve LizardMeasure data structure
using a map of lists (where the path is the key and the measures, with
metric, is the values) to represent measures from the lizard report will
to tightly couple the SonarQube values to the report parsing, making it
hard to work with.

instead the `LizardMeasure` should include number of functions, file
complexity, as well as the path to the file.